removed all the old stuff that was no-longer used after the test changes.
